projects/techcorp_sim/services/api_gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pyjwt==2.9.0
prometheus-client==0.21.0
